+++ src/scripts/chat-message-response.js
@@ -1,13 +1,7 @@
-import { replaceAlias } from "./chat-message-parsing.js";
+import { replaceAlias } from "./chat-message-request.js";
 import { generateResponse } from "./llm.js";
 
 const unkennyResponseFlag = "#UnKennyResponseChatDataInJsonFormat: "
-
-function modifyUnkennyChatData(chatData, addressedActor) {
-    let name = addressedActor.name;
-    let alias = addressedActor.getFlag("unkenny", "alias");
-    chatData.content = replaceAlias(chatData.content, alias, name);
-}
 
 async function triggerResponse(actor, request) {
     let response = await generateResponse(actor, request);
@@ -61,4 +55,4 @@ function overwriteChatMessage() {
     CONFIG.ChatMessage.documentClass = UnkennyChatMessage;
 }
 
-export { modifyUnkennyChatData, overwriteChatMessage, postResponse, processUnKennyResponseSource, triggerResponse, unkennyResponseFlag };
+export { overwriteChatMessage, postResponse, processUnKennyResponseSource, triggerResponse, unkennyResponseFlag };
